Results panel in another class,left panel in another class and getSearch button
