+++ src/com/jalasoft/search/gui/MainFileSearch.java
@@ -21,30 +21,21 @@ import java.awt.GridLayout;
 public class MainFileSearch extends JFrame {
 
     private JButton searchButton;
-    private JButton normalSearchButton;
-    private JButton advancedSearchButton;
-    private JButton videoSearchButton;
-    private JButton musicSearchButton;
-    private JButton recentSearchsButton;
-    private JTextField searchString;
-    private JPanel searchOptionsPanel;
-    private JPanel leftPanel;
     private JPanel northPanel;
-    private JPanel resultsPanel;
-    private JPanel simpleSearchPanel;
-    private JLabel searchLabel;
+
 
     //Class constructor, calls methods to make panels
     public MainFileSearch(String title) {
         super(title);
         makeNorthPanel();
-        makeResultsPanel();
         initializeMainFrame();
     }
 
     /* Initializes main frame by adding all panels in it and set properties for window*/
     private void initializeMainFrame(){
         JFrame mainPanel = new JFrame();
+        ResultsPanel resultsPanel = new ResultsPanel();
+        resultsPanel.add(resultsPanel.getResultsPanel());
         mainPanel.setLayout(new BorderLayout());
         mainPanel.add(northPanel,BorderLayout.NORTH);
         mainPanel.add(resultsPanel,BorderLayout.CENTER );
@@ -65,35 +56,14 @@ public class MainFileSearch extends JFrame {
         simpleSearchPanel.add(simpleSearchPanel.getSimpleSearchPrincipalPanel());
         LeftPanel leftPanel = new LeftPanel();
         leftPanel.add(leftPanel.getLeftPanel());
-        //makeLeftPanel();
         northPanel.add(simpleSearchPanel,BorderLayout.CENTER);
         northPanel.add(leftPanel,BorderLayout.WEST);
         northPanel.add(searchButton,BorderLayout.EAST);
         northPanel.setVisible(true);
     }
-    /*
-    //Initializes left panel that contains buttons to select what kind of search want you use
-    private void makeLeftPanel() {
-        leftPanel = new JPanel();
-        normalSearchButton = new JButton("Simple");
-        advancedSearchButton = new JButton("Advanced");
-        videoSearchButton = new JButton("Video");
-        musicSearchButton = new JButton("Music");
-        recentSearchsButton = new JButton("Recent Searchs");
-        leftPanel.setLayout(new GridLayout(5,1));
-        leftPanel.setBackground(Color.darkGray);
-        leftPanel.add(normalSearchButton);
-        leftPanel.add(advancedSearchButton);
-        leftPanel.add(videoSearchButton);
-        leftPanel.add(musicSearchButton);
-        leftPanel.add(recentSearchsButton);
-    }*/
 
-    /* Initializes result panel*/
-    private void makeResultsPanel() {
-        resultsPanel = new JPanel();
-        resultsPanel.setLayout(new BorderLayout());
-        resultsPanel.setBackground(Color.lightGray);
+    public JButton getSearchButton(){
+        return searchButton;
     }
 
     public static void main(String[] args) {
